Still not really working

+++ display.h
@@ -15,5 +15,5 @@ void display_rec(int rec_bar, unsigned short potvalue, unsigned short *recsample
 void display_clear();
 void display_text(char *text, int line, bool highlight);
 
-void display_adsr(int a, int d, float s, int r, bool select_a, bool select_d, bool select_s, bool select_r);
+void display_adsr(int a, int d, int s, int r, bool select_a, bool select_d, bool select_s, bool select_r);
 #endif
